requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
matplotlib==3.8.2
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
